Add: arbitary baudrate for MCP2518FD, and MCP2515 compatible definitions

+++ src/mcp2515_can.cpp
@@ -931,9 +931,9 @@ byte mcp2515_can::mcp2515_getNextFreeTXBuf(byte* txbuf_n) {               // get
 ** Function name:           begin
 ** Descriptions:            init can and set speed
 *********************************************************************************************************/
-byte mcp2515_can::begin(byte speedset, const byte clockset) {
+byte mcp2515_can::begin(uint32_t speedset, const byte clockset) {
     pSPI->begin();
-    byte res = mcp2515_init(speedset, clockset);
+    byte res = mcp2515_init((byte)speedset, clockset);
 
     return ((res == MCP2515_OK) ? CAN_OK : CAN_FAILINIT);
 }
